Refactor ConnectionRouter into seperate file.
