+++ vpr/src/base/vpr_context.h
@@ -11,6 +11,7 @@
 #include "clustered_netlist.h"
 #include "rr_graph_storage.h"
 #include "rr_node.h"
+#include "rr_rc_data.h"
 #include "tatum/TimingGraph.hpp"
 #include "tatum/TimingConstraints.hpp"
 #include "power.h"
