+++ vpr/src/route/router_delay_profiling.cpp
@@ -27,7 +27,7 @@ bool RouterDelayProfiler::calculate_delay(int source_node, int sink_node, const 
     auto& route_ctx = g_vpr_ctx.routing();
 
     t_rt_node* rt_root = setup_routing_resources_no_net(source_node);
-    enable_router_debug(router_opts, ClusterNetId(), sink_node, 0);
+    enable_router_debug(router_opts, ClusterNetId(), sink_node, 0, &router_);
 
     /* Update base costs according to fanout and criticality rules */
     update_rr_base_costs(1);
